make prevent-invalid-input independent of validation
Previously it's difficult to use prevent-invalid-input and
validation together, because often you want to validate the
complete value but want to prevent invalid input while the
value is incomplete. This commit adds a `allowed-pattern`
attribute that lets you define a different pattern to match
new input with.

Support for prevent setting invalid input programmatically
is dropped. Setting value manually when it's invalid is
too error-prone.